handwriting-recognition-app/requirements.txt
torch
torchvision
opencv-python
matplotlib
numpy
Pillow
scikit-learn
h5py
tqdm
transformers